refactor: reverse urls
