+++ smarter/smarter/apps/dashboard/views/logs/names.py
@@ -2,10 +2,9 @@
 URLs for the logs views.
 """
 
-from smarter.apps.dashboard.const import namespace as dashboard_namespace
 from smarter.common.utils import camel_case_object_name
 
-from .const import namespace as logs_namespace
+from .const import namespace
 from .consumers import RedisLogConsumer
 from .reactapp import TerminalEmulatorLogView
 from .streams import stream_user_logs
@@ -16,7 +15,7 @@ class DashboardLogsReverseNames:
     A class to hold the names of the logs views for easy reference throughout the codebase.
     """
 
-    namespace = ":".join([dashboard_namespace, logs_namespace])
+    namespace = namespace
 
     logs = camel_case_object_name(TerminalEmulatorLogView)
     stream = camel_case_object_name(stream_user_logs)
