+++ smarter/smarter/apps/dashboard/views/passthrough/view.py
@@ -51,7 +51,9 @@ class PromptPassthroughView(SmarterAuthenticatedNeverCachedWebView):
         which only receives JSON dicts of LLM prompts.
         """
         # pylint: disable=C0415
-        from smarter.apps.dashboard.urls.passthrough import PassthroughReverseNames
+        from smarter.apps.dashboard.views.passthrough.urls import (
+            PassthroughReverseNames,
+        )
 
         api_path = reverse(
             ":".join([PromptAPINamespace.namespace, PromptAPINamespace.passthrough]),
