+++ smarter/smarter/apps/dashboard/views/profile/const.py
@@ -1,0 +1,5 @@
+"""
+Profile view constants.
+"""
+
+namespace = "profile"
